+++ turtlebot3_home_service_challenge_tools/CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog for package turtlebot3_home_service_challenge_tools
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.2 (2025-04-04)
+------------------
+* Support for actual robot
+* Updated [actual / simulation] selection with execution arguments.
+* Updated ArUco marker size statement selection with execution arguments.
+* Contributors: ChanHyeong Lee
+
 1.0.1 (2025-03-19)
 ------------------
 * Support for ROS 2 Humble version